pyproject.toml
[tool.poetry]
name = "mangadex-downloader"
version = "0.1.0"
description = "A Python package for downloading manga from MangaDex"
authors = ["Jeffery Fang <Jefferyfang324@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
python-dotenv = "^1.0.1"
aiohttp = {extras = ["speedups"], version = "^3.10.5"}
datetime = "^5.5"
pillow = "^10.4.0"

[tool.poetry.scripts]
mangadex-downloader = "mangadex_downloader.main:main"

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.3"
coverage = "^7.6.1"


[tool.poetry.group.windows.dependencies]
windows-curses = "^2.3.3"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
